the calm before the storm
This commit adds new ID types to the crate, along with a little
plumbing. The intent here is to get rid of the StateID trait and
remove everything that's generic over state IDs and just use u32 for
everything everywhere.

It's kind of sad, but I think the StateID stuff wasn't really pulling
its weight. It was making the implementation quite a bit complex
and safety conditions became more difficult to reason about. And in
particular, we were in this weird tweener state where PatternID was
fixed to `u32` but StateID wasn't. I tried making PatternID generic to
resolve that, but just couldn't do it because it infected everything
(like the `Match` types).

Overall, u32 is really the sweet spot. u8 is generally too small for
most things. u16 is plausible to use in some cases, but it doesn't take
much to blow its budget, particularly for DFAs. u32 uses double the
storage as a u16, but doubling a small number is still pretty small. By
the time you start going beyond what a u32 can store, your memory and
time requirements are so large as to be impractical, so a u64 is never
really needed in practice.

I was tempted to just use u32 everywhere instead of defining wrapper
types, but I decided to be ambitious and see if we can get away with
something a bit tigher than just throwing u32s around. The
'repr(transparent)' bit on the wrapper types is critical so that we can
use StateID/PatternID as if it were a u32, particularly when
transmuting/casting.