make ID iteration nicer
It turns out that we do "slice.iter().enumerate()" a lot, where slice is
*indexed* by a pattern or state ID. But enumerate just gives us a
`usize`. We want a `PatternID` or a `StateID`. It turns out, we almost
always do this for ExactSizeIterator, so we can check the iterator
length at construction and assert that it is correct to build such an
iterator.